implements of the update create report for frontend

+++ src/main/java/com/uniquindio/reporte/model/DTO/location/LocationDTO.java
@@ -1,7 +1,7 @@
 package com.uniquindio.reporte.model.DTO.location;
 
-public record LocationDTO(float latitude,
-                          float longitude,
+public record LocationDTO(Double latitude,
+                          Double longitude,
                           String name,
                           String description) {
 }
